requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
azure-cognitiveservices-speech
azure-storage-blob
opencv-python
mediapipe
textblob
redis
transformers>=4.0.0
simpleaudio
python-dotenv
nltk
pydub
numpy
torch
pyserial
soundfile 
unidecode 
scipy
sounddevice
webrtcvad 
sse_starlette